requirements/base.txt
pygithub3==0.5
python-redmine==0.8.2
